Move ScanSpec to dwio/common.
