+++ velox/dwio/common/CMakeLists.txt
@@ -25,6 +25,7 @@ add_library(
   IoStatistics.cpp
   MemoryInputStream.cpp
   Options.cpp
+  ScanSpec.cpp
   TypeWithId.cpp)
 
 target_link_libraries(
